# tmp_pricelabs_swagger_pretty.json
{
  "swagger": "2.0",
  "info": {
...
    "version": "1.0.0",
    "title": "PriceLabs API Integration",
    "contact": {
      "email": "support@pricelabs.co"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "api.pricelabs.co",
  "basePath": "/v1/integration/api",
...
  "schemes": [
    "https"
  ],
  "security": [
    {
      "integration_token": []
    },
    {
      "integration_name": []
    }
  ],
  "paths": {
...
    "/integration": {
      "post": {
        "tags": [
          "PMS"
        ],
        "summary": "Update integration information",
        "description": "This endpoint can update integration specific details such as refreshing the API token if needed, or adjusting information that allows Pricelabs to send notices/updates.\n\n<b><i>sync_url</i></b>: Endpoint where Pricelabs can send a pricing data to be synced\n<b><i>calendar_trigger_url</i></b>: Endpoint where PriceLabs requests complete listing calendar information\n<b><i>hook_url</i></b>: Endpoint where PriceLabs can send error notifications.\n\nUpdating these three url endpoints results in a test body being sent with only the json body { verify => true } and no headers and requires a 200 response in return in order to be stored.\n\n When you call <b>/integration</b>, please note that if <b><i>regenerate_token</i></b> is true, you will receive a new token in response which should be used to make subsequent API calls.\n\n<b>Delta Push</b>: If you wish to receive only delta updates for prices and restrictions, then you can do so by setting <b><i>delta_only</i></b> flag as true. Once the flag is set to true, you will start to receive only those prices and restrictions which have changed since the last sync. If at any point, you wish to pull full data, you can do so by calling the /get_prices endpoint with the attribute <b><i>delta_only</i></b> as true.\n\n<b>IMPORTANT</b>: The <i>features</i> array should be used to configure only those settings that the PMS intends to actively read from PriceLabs for all connected listings. \n\n  <i>Please note that if a setting is marked as <b>true</b>, it becomes visible to users in PriceLabs, who may enable it with the expectation that the PMS will consume and apply it. Therefore, make sure to enable only those settings that are actually supported and intended to be consumed from PriceLabs.</i>\n  \n  <b>Once the integration is live, any updates to the supported features must first be discussed with PriceLabs. This allows us to coordinate the release of new features to all users and ensure they function as expected.</b>\n\n\n\n",
        "operationId": "integration",
        "parameters": [
          {
            "in": "body",
            "name": "integration",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "integration": {
                  "$ref": "#/definitions/IntegrationData"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Request was processed and returns a message holding new API if requested and successful, or a simple Success and an error message which is populated on errors. <i>Please note that this response will not contain the signed headers as with the other API call responses.</i>",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid integration API key value",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/reservations": {
      "post": {
        "tags": [
          "PMS"
        ],
        "summary": "Update property's reservations",
        "description": "This API will be used to keep the reservations for the listing up to date. Note that the start date and end date are synonymous with check in and check out dates for this API and that total_cost is the complete cost including taxes and fees. Note that listing_id should always be sent as a string, eg., \"12345\".\n<b>IMPORTANT</b> Reservation IDs must be unique across all listings in the PMS. If they are not unique, please append a unique identifier (such as the listing ID) to the reservation ID before sending it to us and make sure to store this updated ID on your end for all future modifications to the same reservation.\nKindly note that once you implement this endpoint, you must push all reservations, including past(since 2020-01-01) and future reservations, to PriceLabs",
        "operationId": "reservations",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "reservations",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "reservations": {
                  "$ref": "#/definitions/Reservations"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was processed, and response contains successes and failures of individually edited reservations\nSuccess will contain only an array of reservation_ids\nFailure contains an error string describing the problem which halted processing, along with the item it was relevant to",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "description": "reservation id processed without error"
                  }
                },
                "failure": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "details": {
                        "$ref": "#/definitions/Reservations"
                      },
                      "error": {
                        "type": "string",
                        "description": "Error message identifying why processing halted"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid integration API key value",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "integration_token": {
      "type": "apiKey",
      "name": "X-INTEGRATION-TOKEN",
      "in": "header",
      "description": "This is an api token. This is generated and provided by PriceLabs."
    },
    "integration_name": {
      "type": "apiKey",
      "name": "X-INTEGRATION-NAME",
      "in": "header",
      "description": "This is a name field for the pms. The name must be approved and provided by PriceLabs."
    }
  },
  "definitions": {
    "Reservations": {
      "type": "array",
      "items": {
        "type": "object",
        "required": [
          "listing_id"
        ],
        "properties": {
          "listing_id": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Reservation"
            }
          }
        }
      }
    },
    "Reservation": {
      "type": "object",
      "required": [
        "reservation_id",
        "currency",
        "start_date",
        "end_date",
        "booked_time",
        "status",
        "total_cost",
        "total_days"
      ],
      "properties": {
        "reservation_id": {
          "type": "string",
          "description": "Unique Reservation ID, should be unique across all listings in the PMS."
        },
        "cancel_time": {
          "type": "string",
          "description": "Should be sent for cancelled bookings. YYYY-MM-DD format"
        },
        "start_date": {
          "type": "string",
          "description": "YYYY-MM-DD format"
        },
        "end_date": {
          "type": "string",
          "description": "YYYY-MM-DD format"
        },
        "booked_time": {
          "type": "string",
          "description": "YYYY-MM-DD format"
        },
        "total_days": {
          "type": "integer"
        },
        "total_cost": {
          "description": "Cost with taxes and fees",
          "type": "number"
        },
        "total_fees": {
          "type": "number",
          "description": "This should be the sum of all fees being charged for the reservation, such as cleaning fees, pet fees, extra person fees, parking fees, etc."
        },
        "total_taxes": {
          "type": "number",
          "description": "This should be the sum of all taxes charged for the reservation, including VAT, city tax, etc"
        },
        "host_payout": {
          "type": "number",
          "description": "This is the amount that a property manager will receive after commission deductions. Essentially, it is the revenue that a host earns from a reservation."
        },
        "ota_commission": {
          "type": "number",
          "description": "This is the commission charged by the platform for reservations."
        },
        "rental_revenue": {
          "type": "number",
          "description": "This is the base nightly amount of a reservation, excluding any fees and taxes (important)."
        },
        "currency": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "description": "Reservation Statuses. booked: Confirmed Reservations, blocked -  maintenance or owner blocks, unconfirmed - unconfirmed reservations, canceled - cancelled reservations",
          "enum": [
            "booked",
            "blocked",
            "unconfirmed",
            "canceled"
          ]
        }
      }
    },
...
    "IntegrationData": {
      "type": "object",
      "required": [
        "sync_url",
        "calendar_trigger_url",
        "hook_url",
        "regenerate_token"
      ],
      "properties": {
        "sync_url": {
          "type": "string"
        },
        "calendar_trigger_url": {
          "type": "string"
        },
        "regenerate_token": {
          "type": "boolean",
          "default": true,
          "description": "If not sent, it will not regenerate token. If sent blank, it will regenerate the token by default."
        },
        "hook_url": {
          "type": "string"
        },
        "features": {
          "$ref": "#/definitions/FeaturesControl"
        }
      }
    },
...
    "FeaturesControl": {
      "type": "object",
      "properties": {
        "min_stay": {
          "type": "boolean"
        },
        "check_in": {
          "type": "boolean"
        },
        "check_out": {
          "type": "boolean"
        },
        "monthly_weekly_discounts": {
          "type": "boolean"
        },
        "extra_person_fee": {
          "type": "boolean"
        },
        "los_pricing": {
          "type": "boolean"
        },
        "delta_only": {
          "type": "boolean",
          "description": "If this flag is set to true, PriceLabs will push only delta to the sync_url. In other words, only those prices and restrictions will be pushed which have changed since the last sync. If this flag is set to false, then all prices and restrictions will be pushed."
        }
      }
    },
...
  }
}